CLN: refactoring of routines into src modules
